# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "EHPS API",
    "description" : "Equipment Health & Performance System — semiconductor fab equipment maintenance. Covers fab-coordinator machine/technician management, routine checkups, alert escalation and approval, repair assignment/completion, and the equipment health dashboard. All endpoints (except /api/auth/**) require a bearer JWT obtained from POST /api/auth/login, and are scoped by role: FAB_COORDINATOR, MANAGER, or TECHNICIAN. See docs/API-CONTRACT.md for the full endpoint contract.",
    "version" : "2.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
...
    "/api/manager/alerts/{alertId}/rejection" : {
      "post" : {
        "tags" : [ "Manager — Alerts" ],
        "summary" : "Reject an alert",
        "description" : "Rejects an escalated alert. Requires role MANAGER.",
        "operationId" : "rejectAlert",
        "parameters" : [ {
          "name" : "alertId",
          "in" : "path",
          "description" : "Alert ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAlertActionResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/manager/alerts/{alertId}/approval" : {
      "post" : {
        "tags" : [ "Manager — Alerts" ],
        "summary" : "Approve an alert",
        "description" : "Approves an escalated alert, allowing it to proceed to repair assignment. Requires role MANAGER.",
        "operationId" : "approveAlert",
        "parameters" : [ {
          "name" : "alertId",
          "in" : "path",
          "description" : "Alert ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAlertActionResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/fab/alerts/{alertId}/repair-assignment" : {
      "post" : {
        "tags" : [ "Fab — Alerts" ],
        "summary" : "Assign a repair technician to an alert",
        "description" : "Assigns a technician to repair the machine associated with an approved alert. Requires role FAB_COORDINATOR.",
        "operationId" : "assignRepairTechnician",
        "parameters" : [ {
          "name" : "alertId",
          "in" : "path",
          "description" : "Alert ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AssignRepairRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAssignRepairResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/fab/alerts/{alertId}/escalation" : {
      "post" : {
        "tags" : [ "Fab — Alerts" ],
        "summary" : "Escalate an alert to manager",
        "description" : "Sends a pending alert to the manager for approval. Requires role FAB_COORDINATOR.",
        "operationId" : "sendToManager",
        "parameters" : [ {
          "name" : "alertId",
          "in" : "path",
          "description" : "Alert ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseEscalationResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/auth/register" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Register a new user",
        "description" : "Creates a new user account with a role (technician, manager, or fab_coordinator). Public endpoint, no authentication required. Returns the created user without a token.",
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseUserResponse"
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
    "/api/auth/login" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Log in and obtain a JWT",
        "description" : "Authenticates a user by email/password and returns a bearer JWT plus the user profile. Public endpoint, no authentication required.",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseAuthResponse"
                }
              }
            }
          }
        },
        "security" : [ ]
      }
    },
...
    "/api/technician/repairs/alerts" : {
      "get" : {
        "tags" : [ "Technician — Repairs" ],
        "summary" : "List my approved repair alerts",
        "description" : "Returns approved alerts assigned to the calling technician for repair. Requires role TECHNICIAN.",
        "operationId" : "getApprovedRepairAlerts",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListApprovedRepairAlertResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/technician/repairs/alerts/{alertId}" : {
      "get" : {
        "tags" : [ "Technician — Repairs" ],
        "summary" : "Get a single approved repair alert",
        "description" : "Returns a single approved repair alert assigned to the calling technician. Requires role TECHNICIAN.",
        "operationId" : "getApprovedRepairAlert",
        "parameters" : [ {
          "name" : "alertId",
          "in" : "path",
          "description" : "Alert ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseApprovedRepairAlertResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/manager/alerts" : {
      "get" : {
        "tags" : [ "Manager — Alerts" ],
        "summary" : "List all alerts",
        "description" : "Returns every alert visible to managers. Requires role MANAGER.",
        "operationId" : "getAllAlerts",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAlertResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/manager/alerts/pending-approval" : {
      "get" : {
        "tags" : [ "Manager — Alerts" ],
        "summary" : "List alerts pending manager approval",
        "description" : "Returns alerts escalated by a fab coordinator that are awaiting manager approval. Requires role MANAGER.",
        "operationId" : "getPendingManagerApprovalAlerts",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAlertResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/fab/alerts/{alertId}/candidate-technicians" : {
      "get" : {
        "tags" : [ "Fab — Alerts" ],
        "summary" : "List candidate technicians for an alert",
        "description" : "Returns technicians qualified to repair the machine associated with the given alert. Requires role FAB_COORDINATOR.",
        "operationId" : "getMatchingTechnicians",
        "parameters" : [ {
          "name" : "alertId",
          "in" : "path",
          "description" : "Alert ID",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTechnicianResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/fab/alerts/pending" : {
      "get" : {
        "tags" : [ "Fab — Alerts" ],
        "summary" : "List pending alerts",
        "description" : "Returns alerts awaiting fab-coordinator triage. Requires role FAB_COORDINATOR.",
        "operationId" : "getPendingAlerts",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAlertResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/fab/alerts/approved-unassigned" : {
      "get" : {
        "tags" : [ "Fab — Alerts" ],
        "summary" : "List approved, unassigned alerts",
        "description" : "Returns manager-approved alerts that do not yet have a repair technician assigned. Requires role FAB_COORDINATOR.",
        "operationId" : "getApprovedUnassignedAlerts",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListAlertResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TechnicianResponse" : {
        "type" : "object",
        "properties" : {
          "technicianId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "speciality" : {
            "type" : "string"
          }
        }
      },
...
      "AlertActionResponse" : {
        "type" : "object",
        "properties" : {
          "alertId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "updatedStatus" : {
            "type" : "string"
          },
          "managerId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ApiResponseAlertActionResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AlertActionResponse"
          }
        }
      },
...
      "AssignRepairRequest" : {
        "type" : "object",
        "properties" : {
          "technicianId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "repairDate" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "technicianId" ]
      },
      "ApiResponseAssignRepairResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AssignRepairResponse"
          }
        }
      },
      "AssignRepairResponse" : {
        "type" : "object",
        "properties" : {
          "alertId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "technicianId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "workType" : {
            "type" : "string"
          },
          "workDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "ApiResponseEscalationResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/EscalationResponse"
          }
        }
      },
      "EscalationResponse" : {
        "type" : "object",
        "properties" : {
          "alertId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "updatedStatus" : {
            "type" : "string"
          }
        }
      },
      "RegisterRequest" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "minLength" : 1
          },
          "email" : {
            "type" : "string",
            "pattern" : "^[A-Za-z0-9._%+-]+@ehps\\.com$"
          },
          "phone" : {
            "type" : "string",
            "pattern" : "^\\d{10}$"
          },
          "password" : {
            "type" : "string",
            "pattern" : "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "manager", "fab_coordinator", "technician" ]
          },
          "speciality" : {
            "type" : "string"
          }
        },
        "required" : [ "name", "role" ]
      },
      "ApiResponseUserResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/UserResponse"
          }
        }
      },
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "empId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "speciality" : {
            "type" : "string"
          }
        }
      },
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "format" : "email"
          },
          "password" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "password" ]
      },
      "ApiResponseAuthResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/AuthResponse"
          }
        }
      },
      "AuthResponse" : {
        "type" : "object",
        "properties" : {
          "token" : {
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserResponse"
          }
        }
      },
...
      "ApiResponseListApprovedRepairAlertResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ApprovedRepairAlertResponse"
            }
          }
        }
      },
      "ApprovedRepairAlertResponse" : {
        "type" : "object",
        "properties" : {
          "alertId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "problemMeasure" : {
            "type" : "string"
          },
          "severity" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "raisedOn" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "ApiResponseApprovedRepairAlertResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/ApprovedRepairAlertResponse"
          }
        }
      },
...
      "ApiResponseListTechnicianResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TechnicianResponse"
            }
          }
        }
      },
...
      "AlertResponse" : {
        "type" : "object",
        "properties" : {
          "alertId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "machineCode" : {
            "type" : "string"
          },
          "problemMeasure" : {
            "type" : "string"
          },
          "severity" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "raisedOn" : {
            "type" : "string",
            "format" : "date"
          },
          "fabCoordinatorId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "approvedById" : {
            "type" : "integer",
            "format" : "int64"
          },
          "assignedTechnicianId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ApiResponseListAlertResponse" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AlertResponse"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}